+++ to-do/src/main/java/com/practica/todo/repositorios/Usuariosrep.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.practica.todo.entidades.*;
 
 @Repository
-public interface Usuariosrep extends JpaRepository  <Usuario, Long>{
+public interface Usuariosrep extends JpaRepository  <Usuario, Integer>{
 
     Optional<Usuario> findByEmail(String email);
